Refactor MIMIC mortality examples into notebooks
